feat: 32-bit (ILP32) support for arm and 386
Enable building and running on 32-bit targets now that go-audio-capture v0.5.0
ships architecture-aware ALSA layouts and a guard for unsupported arches.

The only 32-bit compile blocker here was internal/sysinfo/collect_linux.go:
syscall.Statfs_t.Bsize is int32 on ILP32 but int64 on LP64, so widen it before
the block-count multiply.

Also add build/CI coverage: a build-arm Taskfile target (GOARM=6 for ARMv6
reach), go vet across amd64/arm64/arm/386, a linux/arm CI build step plus arm
and 386 in the vet matrix, and README updates documenting 32-bit support.

Verified: builds for arm (GOARM=6/7), 386, amd64, arm64; go vet clean on all
four; race tests pass. Hardware-validated with a 32-bit ARM binary capturing
from a real USB card and streaming Opus over RTSP end to end.

+++ internal/sysinfo/collect_linux.go
@@ -169,7 +169,9 @@ func diskUsage(path string) (total, used int64, ok bool) {
 	// used = Blocks - Bfree (both include root-reserved space) so the figure
 	// matches df's Used; Bavail would exclude reserved blocks from free only,
 	// overstating used by the reserved amount.
-	bsize := st.Bsize
+	// Bsize is int64 on LP64 but int32 on ILP32 (arm, 386); widen once so the
+	// multiplications below are int64 on every architecture.
+	bsize := int64(st.Bsize)
 	total = int64(st.Blocks) * bsize //nolint:gosec // block counts fit int64 on real filesystems
 	free := int64(st.Bfree) * bsize  //nolint:gosec // block counts fit int64 on real filesystems
 	used = total - free
